moved the code in class functions/console into the classes/console
